% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cupfunc.R
\name{add_ref_annotations}
\alias{add_ref_annotations}
\title{\code{add_ref_annotations} integrate GWAS summary data with support file}
\usage{
add_ref_annotations(ss, DT)
}
\arguments{
\item{ss}{data.table object for snp manifest}

\item{DT}{data.table containing GWAS summary stats}
}
\value{
data.table object
}
\description{
\code{add_ref_annotations} integrate GWAS summary data with support file
}
