% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_alleles.R
\name{align_alleles}
\alias{align_alleles}
\title{\code{align_alleles} compute variance weighted euclidean distance between two PC loading vectors a and b}
\usage{
align_alleles(gwas.DT, ref.DT, check = FALSE)
}
\arguments{
\item{gwas.DT}{a data.table of GWAS data see \code{\link{get_gwas_data}}}

\item{ref.DT}{a data.table of reference alleles (see details)}

\item{check}{boolean scalar whether to add column of transformations for checking}
}
\value{
gwas.DT with allele aligned DT
}
\description{
\code{align_alleles} compute variance weighted euclidean distance between two PC loading vectors a and b
}
\details{
this routine lines up on reference `a1` this is arbitrary but as long as consistency is maintained
then the basis can be correctly created
}
