% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_alleles.R
\name{ambiguous}
\alias{ambiguous}
\title{\code{ambiguous} get ambiguous variants}
\usage{
ambiguous(a1, a2)
}
\arguments{
\item{a1}{a vector of alleles AGCT}

\item{a2}{a vector of alleles AGCT}
}
\value{
a boolean vector
}
\description{
\code{ambiguous} get ambiguous variants
}
