% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cupfunc.R
\name{basis_pp}
\alias{basis_pp}
\title{This function computes the posterior prob that a SNP is causal in a set of traits
\code{basis_pp} computes the posterior probability that a SNP is causal across a set of traits}
\usage{
basis_pp(BF, emp_pi)
}
\arguments{
\item{bf}{a vector of approximate Bayes Factors using Wakefield's method.}

\item{a}{scalar or vector of posterior probabilites}
}
\value{
pi a scalar empirical prior
}
\description{
This function computes the posterior prob that a SNP is causal in a set of traits
\code{basis_pp} computes the posterior probability that a SNP is causal across a set of traits
}
