% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cupfunc.R
\name{bayesian_shrinkage}
\alias{bayesian_shrinkage}
\title{This function computes an approximate Bayes Factor for whether a given SNP is causal
in a set of basis traits assuming a single causal variant.
\code{bayesian_shrinkage} computes an approximate Bayes Factor for whether a given SNP is causal
in a set of basis traits assuming a single causal variant. Using an emprically derived
prior computes a posterior probability.}
\usage{
bayesian_shrinkage(DT, tquant = 0.9999)
}
\arguments{
\item{DT}{basis data.table object}

\item{tquant}{a scalar representing quantile on which to truncate infinite Bayes Factors (DEFAULT 0.9999)}
}
\value{
data.table object
}
\description{
This function computes an approximate Bayes Factor for whether a given SNP is causal
in a set of basis traits assuming a single causal variant.
\code{bayesian_shrinkage} computes an approximate Bayes Factor for whether a given SNP is causal
in a set of basis traits assuming a single causal variant. Using an emprically derived
prior computes a posterior probability.
}
