% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cupfunc.R
\name{ca}
\alias{ca}
\title{This function computes an estimate of allele count of unexposed controls
\code{ca} estimate of allele count of unexposed controls}
\usage{
ca(n0, f)
}
\arguments{
\item{n0}{a vector or scalar of number of control samples}

\item{f}{a vector of reference allele frequencies}
}
\value{
a numeric vector
}
\description{
This function computes an estimate of allele count of unexposed controls
\code{ca} estimate of allele count of unexposed controls
}
