% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cupfunc.R
\name{cc}
\alias{cc}
\title{This function computes an estimate of allele count of  unexposed cases
\code{cc} estimate of allele count of unexposed cases}
\usage{
cc(n1, a, b, theta)
}
\arguments{
\item{n1}{a vector or scalar of number of control samples}

\item{a}{a vector of estimates for allele count of  unexposed controls}

\item{b}{a vector of estimates for allele count of  exposed controls}
}
\value{
a numeric vector
see also \code{\link{ca}} and \code{\link{cb}}
}
\description{
This function computes an estimate of allele count of  unexposed cases
\code{cc} estimate of allele count of unexposed cases
}
