% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_alleles.R
\name{comp}
\alias{comp}
\title{\code{comp} get complement for a set of variants}
\usage{
comp(cv)
}
\arguments{
\item{cv}{a vector of variants}
}
\value{
a vector of complement variants
}
\description{
\code{comp} get complement for a set of variants
}
