% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{compute_proj_var}
\alias{compute_proj_var}
\title{analytically compute the variance of a projection given a reference set of genotypes
\code{compute_proj_var}}
\usage{
compute_proj_var(DT, w.DT, shrink.DT, ref_gt_dir, method = "ws_emp",
  quiet = TRUE)
}
\arguments{
\item{DT}{a data.table - as returned by \code{\link{get_gwas_data}}}

\item{w.DT}{a data.table - a data table of weights/rotations from a  basis first column is 'pid' and subsequent columns for each principal component.}

\item{shrink.DT}{a data.table - as returned by \code{\link{compute_shrinkage_metrics}}}

\item{ref_gt_dir}{scalar - path to a dir of R objects named CHR_1kg.RData containing reference GT in snpMatrix format}

\item{method}{scalar - shrinkage method to use (default ws_emp)}

\item{quiet}{a scalar - boolean whether to show progress messages}
}
\value{
a scalar of variances for each principal component
}
\description{
analytically compute the variance of a projection given a reference set of genotypes
\code{compute_proj_var}
}
