% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance.R
\name{compute_seb_proj_var}
\alias{compute_seb_proj_var}
\title{analytically compute the variance of a projection given a standard error of beta for a given trait
\code{compute_seb_proj_var}}
\usage{
compute_seb_proj_var(gwas.DT, shrink.DT, man.DT, w.DT, ref_gt_dir,
  method = "shrinkage", quiet = FALSE)
}
\arguments{
\item{gwas.DT}{a data.table - variants and summary statistics}

\item{shrink.DT}{a data.table - as returned by \code{\link{compute_shrinkage_metrics}}}

\item{man.DT}{a data.table - snp manifest file that maps snps to ld.block}

\item{w.DT}{a data.table - a data table of weights/rotations from a  basis first column is 'pid' and subsequent columns for each principal component.}

\item{ref_gt_dir}{scalar - path to a dir of R objects named CHR_1kg.RData containing reference GT in snpMatrix format}

\item{method}{scalar - shrinkage method to use (default shrinkage) but also valid is none}

\item{quiet}{a scalar - boolean whether to show progress messages}
}
\value{
a scalar of variances for each principal component
}
\description{
analytically compute the variance of a projection given a standard error of beta for a given trait
\code{compute_seb_proj_var}
}
