% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance.R
\name{compute_seb_proj_var_sparse}
\alias{compute_seb_proj_var_sparse}
\title{analytically compute the variance of a projection given a standard error of beta for a given trait for sparse basis
\code{compute_seb_proj_var_sparse}}
\usage{
compute_seb_proj_var_sparse(
  gwas.DT,
  shrink.DT,
  w.DT,
  sm,
  method = "shrinkage",
  quiet = FALSE
)
}
\arguments{
\item{gwas.DT}{a data.table - variants and summary statistics}

\item{shrink.DT}{a data.table - as returned by \code{\link{compute_shrinkage_metrics}}}

\item{w.DT}{a data.table - a data table of weights/rotations from a  basis first column is 'pid' and subsequent columns for each principal component.}

\item{sm}{snpMatrix object - a snp matrix object of reference genotypes for}

\item{method}{scalar - shrinkage method to use (default shrinkage) but also valid is none}

\item{quiet}{a scalar - boolean whether to show progress messages}
}
\value{
a scalar of variances for each principal component
}
\description{
analytically compute the variance of a projection given a standard error of beta for a given trait for sparse basis
\code{compute_seb_proj_var_sparse}
}
