% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_lor.R
\name{control_prior_shape}
\alias{control_prior_shape}
\title{This function computes shape parameters for beta distribution matching a binomial proportion B(n,f)/n
\code{control_prior_shape} shape parameters for beta distribution matching a binomial proportion B(n,f)/n}
\usage{
control_prior_shape(f, n)
}
\arguments{
\item{f}{a scalar - allele frequency in the controls}

\item{n}{a scalar - number if indivudals (not observations) used to estimate f}
}
\value{
list of shape parameters for a beta distributon
\enumerate{
\item a0 - alpha shape parameter in controls
\item b0 - beta shape parameter in controls
}
}
\description{
This function computes shape parameters for beta distribution matching a binomial proportion B(n,f)/n
\code{control_prior_shape} shape parameters for beta distribution matching a binomial proportion B(n,f)/n
}
