% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin2or.R
\name{convertBetaToOR}
\alias{convertBetaToOR}
\title{\code{convertBetaToOR}}
\usage{
convertBetaToOR(thresh, N, b, seb, m)
}
\arguments{
\item{thresh}{- threshold value to use (if missing use mean of Y).}

\item{N}{- Number of samples}

\item{b}{- beta coefficient from linear model fit}

\item{seb}{- standard error of beta coefficient}

\item{m}{- minor allele freq}
}
\value{
a vector of lists containing OR,betas and other intermediates
}
\description{
\code{convertBetaToOR}
}
