% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cupfunc.R
\name{create_basis}
\alias{create_basis}
\title{This function creates a basis
\code{create_basis} creates a trait snp matrix that is suitable for basis creation and projection}
\usage{
create_basis(gwas.DT, shrink.DT, apply.shrinkage = TRUE)
}
\arguments{
\item{gwas.DT}{data.table object for basis traits as returned by \code{\link{get_gwas_data}}}

\item{shrink.DT}{data.table object of matching shrinkage estimates returned by \code{\link{compute_shrinkage_metrics}}}

\item{method}{scalar vector (either shrinkage or none), emp uses empirically generated MAF SE, est uses and estimate.}
}
\value{
a prcomp object representing the basis
}
\description{
This function creates a basis
\code{create_basis} creates a trait snp matrix that is suitable for basis creation and projection
}
