% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cupfunc.R
\name{create_ds_matrix}
\alias{create_ds_matrix}
\title{This function creates a trait snp matrix
\code{create_ts_matrix} creates a trait snp matrix that is suitable for basis creation and projection}
\usage{
create_ds_matrix(bDT, sDT, method = c("emp", "est", "memp", "mest"))
}
\arguments{
\item{bDT}{data.table object for basis traits as returned by \code{\link{get_gwas_data}}}

\item{sDT}{data.table object of matching shrinkage estimates returned by \code{\link{compute_shrinkage_metrics}}}

\item{method}{scalar vector (either emp or est), emp uses empirically generated MAF SE, est uses and estimate.}
}
\value{
a matrix.
}
\description{
This function creates a trait snp matrix
\code{create_ts_matrix} creates a trait snp matrix that is suitable for basis creation and projection
}
