% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_lor.R
\name{e_lor}
\alias{e_lor}
\title{This function computes expected log odds ratio for a set of shape parameters
\code{e_lor} computes expected log odds ratio for a set of shape parameters}
\usage{
e_lor(b1, a0, b0, a1)
}
\arguments{
\item{b1}{a scalar - beta distribution beta shape parameter for prior on allele frequency in cases}

\item{a0}{a scalar - beta distribution alpha shape parameter for prior on allele frequency in controls}

\item{a1}{a scalar - beta distribution alpha shape parameter for prior on allele frequency in cases}

\item{a0}{a scalar - beta distribution beta shape parameter for prior on allele frequency in controls}
}
\value{
scalar - expected log(OR)
}
\description{
This function computes expected log odds ratio for a set of shape parameters
\code{e_lor} computes expected log odds ratio for a set of shape parameters
}
