% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_lor.R
\name{est_a1b1}
\alias{est_a1b1}
\title{This function estimates shape parameters for a prior distribution of control allele frequencies
\code{est_a1b1} estimate shape parameters a1,b1}
\usage{
est_a1b1(a0, b0, target.or, target.prob, nsim)
}
\arguments{
\item{a0}{a scalar - beta distribution alpha shape parameter for prior on allele frequency in controls}

\item{b0}{a scalar - beta distribution beta shape parameter for prior on allele frequency in controls}

\item{target.or}{a scalar - an odds ratio threshold to compute P(sim.or > target.or)}

\item{target.prob}{a scalar - a probability that a sampled variant will exceed target.or}

\item{nsim}{a scalar - number of simulations to use when estimating parameters}
}
\value{
list - shape parameters for beta distribution satisfying target.or,target.prob and control prior distribution constraints.
}
\description{
This function estimates shape parameters for a prior distribution of control allele frequencies
\code{est_a1b1} estimate shape parameters a1,b1
}
