% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_alleles.R
\name{flip_allele}
\alias{flip_allele}
\title{\code{align_allele} given a data.table of pid,a1,a2 get a list of SNPs where alleles are flipped wrt to basis reference}
\usage{
flip_allele(DT, ref.DT)
}
\arguments{
\item{DT}{- a data.table object - columns chr,position,a1 and a2}

\item{ref.DT}{- a data.table of reference alleles (see details)}
}
\value{
an index wrt to DT of OR that require flipping
}
\description{
\code{align_allele} given a data.table of pid,a1,a2 get a list of SNPs where alleles are flipped wrt to basis reference
}
