% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_lor.R
\name{fopt}
\alias{fopt}
\title{This function computes the optimal shape parameter a1 for a given set of constraints
shape parameters estimate the P(OR > target.or).
\code{fopt} estimate the P(OR > target.or) given shape parameters for case and control allele frequency}
\usage{
fopt(a1, target.prob, target.or, p, a0, b0)
}
\arguments{
\item{a1}{a scalar - beta distribution alpha shape parameter for prior on allele frequency in cases}

\item{target.prob}{a scalar - a probability that a sampled variant will exceed target.or}

\item{target.or}{a scalar - an odds ratio threshold to compute P(sim.or > target.or)}

\item{p}{- a vector of values to optimise over}

\item{a0}{a scalar - beta distribution alpha shape parameter for prior on allele frequency in controls}

\item{b0}{a scalar - beta distribution beta shape parameter for prior on allele frequency in controls}
}
\value{
scalar - P(sim.or>target.or)
}
\description{
This function computes the optimal shape parameter a1 for a given set of constraints
shape parameters estimate the P(OR > target.or).
\code{fopt} estimate the P(OR > target.or) given shape parameters for case and control allele frequency
}
