% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance.R
\name{get_seb}
\alias{get_seb}
\title{An internal helper function to get obtain standard error of beta estimates from a GWAS.DT data.frame object
\code{get_seb}}
\usage{
get_seb(gwas.DT)
}
\arguments{
\item{gwas.DT}{a data.table - variants and summary statistics}
}
\value{
a data.table with seb column added if not already present
}
\description{
An internal helper function to get obtain standard error of beta estimates from a GWAS.DT data.frame object
\code{get_seb}
}
