% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cupfunc.R
\name{logsum}
\alias{logsum}
\title{helper function to sum logs without loss of precision
\code{logsum} sums logs without loss of precision}
\usage{
logsum(x)
}
\arguments{
\item{x}{a vector of logs to sum}
}
\value{
a scalar
}
\description{
helper function to sum logs without loss of precision
\code{logsum} sums logs without loss of precision
}
