% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_lor.R
\name{lor_constraint}
\alias{lor_constraint}
\title{This function computes a probability for a given configuration of beta distribution
shape parameters estimate the P(OR > target.or).
\code{lor_constraint} estimate the P(OR > target.or) given shape parameters for case and control allele frequency}
\usage{
lor_constraint(a1, a0, b0, p0, target.or, nsim)
}
\arguments{
\item{a1}{a scalar - beta distribution alpha shape parameter for prior on allele frequency in cases}

\item{a0}{a scalar - beta distribution alpha shape parameter for prior on allele frequency in controls}

\item{b0}{a scalar - beta distribution beta shape parameter for prior on allele frequency in controls}

\item{p0}{a vector - a set of samples from beta(a0,b0)}

\item{target.or}{a scalar - an odds ratio threshold to compute P(sim.or > target.or)}
}
\value{
scalar - P(sim.or>target.or)
}
\description{
This function computes a probability for a given configuration of beta distribution
shape parameters estimate the P(OR > target.or).
\code{lor_constraint} estimate the P(OR > target.or) given shape parameters for case and control allele frequency
}
