% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_lor.R
\name{lor_f}
\alias{lor_f}
\title{This function samples from the posterior distribution of a given control allele frequency, for different genotype configurations
\code{lor_f} sample from log(or) posterior distribution given an allele frequency.}
\usage{
lor_f(f0, n, target.or, target.prob, n.steps)
}
\arguments{
\item{f0}{a scalar - allele frequency in cases to simulate}

\item{n}{a scalar - number of individuals sampled}

\item{target.or}{a scalar - an odds ratio threshold to compute P(sim.or > target.or)}

\item{target.prob}{a scalar - a probability that a sampled variant will exceed target.or}

\item{n.steps}{an integer - length of search grid to employ - larger gives a more accurate integral estimate}
}
\value{
a list object
\enumerate{
\item 00 - mean lor for reference hom
\item 01 - mean lor for het
\item 11 - mean lor for alternative hom
\item quant00 - quantiles for posterior distribution of allele frequency in cases for ref hom
\item quant01 - quantiles for posterior distribution of allele frequency in cases for het
\item quant11 - quantiles for posterior distribution of allele frequency in cases for alt hom
}
}
\description{
This function samples from the posterior distribution of a given control allele frequency, for different genotype configurations
\code{lor_f} sample from log(or) posterior distribution given an allele frequency.
}
