% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lyons_egpa.R
\docType{data}
\name{lyons_egpa}
\alias{lyons_egpa}
\title{EGPA combined GWAS summary statistics}
\format{A data.table object of odds ratios, beta standard errors and p values.}
\usage{
data(lyons_egpa)
}
\description{
Summary data from GWAS study of EGPA of 534 cases and 6,688 controls
}
\examples{
data(lyons_egpa)
head(lyons_egpa)
}
\references{
Lyons et al. Nat Commun 10, 5120 (2019) doi:10.1038/s41467-019-12515-9
(\href{https://doi.org/10.1038/s41467-019-12515-9}{DOI})
}
\keyword{datasets}
