% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cupfunc.R
\name{maf_se_estimate}
\alias{maf_se_estimate}
\title{Compute minor allele frequency shrinkage
\code{maf_se_estimate} computes a shrinkage metric for a given list of minor allele frequencies'}
\usage{
maf_se_estimate(f)
}
\arguments{
\item{f}{a vector of minor allele frequencies taken from some reference population.}
}
\value{
a vector of shrinkage metrics
}
\description{
Compute minor allele frequency shrinkage
\code{maf_se_estimate} computes a shrinkage metric for a given list of minor allele frequencies'
}
