% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cupfunc.R
\name{maf_se_estimate_sample_size}
\alias{maf_se_estimate_sample_size}
\title{Compute minor allele frequency shrinkage using sample size
\code{maf_se_estimate_sample_size} computes component of standard error of beta due to minor allele frequency}
\usage{
maf_se_estimate_sample_size(N, p, theta, f)
}
\arguments{
\item{N}{a vector or scalar of total number od samples}

\item{p}{a vector or scalar of p values}

\item{theta}{a vector or scalar of odds ratios}

\item{f}{a vector of reference allele frequencies}
}
\value{
a numeric vector
}
\description{
Compute minor allele frequency shrinkage using sample size
\code{maf_se_estimate_sample_size} computes component of standard error of beta due to minor allele frequency
}
