% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{mvs_perm}
\alias{mvs_perm}
\title{\code{mvs_perm} sample from a multivariate normal distribution}
\usage{
mvs_perm(m, sigma, n = 2)
}
\arguments{
\item{m}{a vector -  mean values}

\item{sigma}{a matrix  - a positive definite covariance matrix}

\item{n}{a scalar - number of samples to generate (default 2)}
}
\value{
a matrix of sampled values
}
\description{
\code{mvs_perm} sample from a multivariate normal distribution
}
