% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{mvs_sigma}
\alias{mvs_sigma}
\title{Code to compute sigma - genotype covariance matrix
\code{mvs_sigma} create a geneotype convariance matrix}
\usage{
mvs_sigma(r, quiet = TRUE)
}
\arguments{
\item{r}{a matrix of pairwise \eqn{r^2} measures generated by snpStats}

\item{quiet}{boolean whether to return warnings messages default=TRUE}
}
\value{
a positive definite approximation of covariance matrix
}
\description{
Code to compute sigma - genotype covariance matrix
\code{mvs_sigma} create a geneotype convariance matrix
}
