% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin2or.R
\name{n_alt}
\alias{n_alt}
\title{compute counts for alternative (minor) allele within a population
\code{n_alt} compute expected count of alt alleles for a given sample}
\usage{
n_alt(N, m)
}
\arguments{
\item{N}{sample size}

\item{m}{minor allele freq}
}
\value{
a vector of allele counts
}
\description{
compute counts for alternative (minor) allele within a population
\code{n_alt} compute expected count of alt alleles for a given sample
}
