% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin2or.R
\name{n_het}
\alias{n_het}
\title{compute counts for heterozygous allele within a population
\code{n_het} compute expected count of heterozygous\ alleles for a given sample}
\usage{
n_het(N, m)
}
\arguments{
\item{N}{sample size}

\item{m}{minor allele freq}
}
\value{
a vector of allele counts
}
\description{
compute counts for heterozygous allele within a population
\code{n_het} compute expected count of heterozygous\ alleles for a given sample
}
