% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin2or.R
\name{n_ref}
\alias{n_ref}
\title{compute counts for different reference (major) allele within a population
\code{n_ref} compute expected count of ref alleles for a given sample}
\usage{
n_ref(N, m)
}
\arguments{
\item{N}{sample size}

\item{m}{minor allele freq}
}
\value{
a vector of allele counts
}
\description{
compute counts for different reference (major) allele within a population
\code{n_ref} compute expected count of ref alleles for a given sample
}
