% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cupfunc.R
\name{p2z}
\alias{p2z}
\title{convert p value  to a signed Z score
\code{p2z} p value to a signed Z score}
\usage{
p2z(p.val, lor)
}
\arguments{
\item{lor}{a vector of log odds ratios}

\item{p}{a vector of p values}
}
\value{
a vector of signed Z scores
}
\description{
convert p value  to a signed Z score
\code{p2z} p value to a signed Z score
}
