% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{pairwise_euc_matrix}
\alias{pairwise_euc_matrix}
\title{\code{pairwise_euc_matrix} compute a pairwise matrix of variance weighted distances across PC loadings}
\usage{
pairwise_euc_matrix(x, vexp, pcidx)
}
\arguments{
\item{x}{numeric matrix of loadings (e.g pc$x)}

\item{vexp}{a numeric vector of variance explained for a set of principle components (e.g summary(pc)$importance[2,])}

\item{pcidx}{(optional) parameter sepcifying a single principle component to use when computing distance (note no longer variance weighted)}
}
\value{
a numeric matrix
}
\description{
\code{pairwise_euc_matrix} compute a pairwise matrix of variance weighted distances across PC loadings
}
