% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_lor.R
\name{post_lor}
\alias{post_lor}
\title{This function samples from the posterior distribution of log(OR) based on the observation of a case genotype
\code{post_lor} samples from the posterior distribution of log(OR) given genotype.}
\usage{
post_lor(gt = c(0, 1, 2), a1, b1, p0, nsim)
}
\arguments{
\item{gt}{a scalar - takes the values 0 - reference hom, 1 - het, 2 - alternative hom.}

\item{a1}{a scalar - beta distribution alpha shape parameter for prior on allele frequency in cases}

\item{b1}{a scalar - beta distribution beta shape parameter for prior on allele frequency in cases}

\item{p0}{a vector - a set of samples from beta(a0,b0) i.e. prior on control allele frequency}

\item{nsim}{a scalar - number of log odds ratios to simulate from the posterior distribution}
}
\value{
vector of simulated log(OR) from the posterior.
}
\description{
This function samples from the posterior distribution of log(OR) based on the observation of a case genotype
\code{post_lor} samples from the posterior distribution of log(OR) given genotype.
}
