% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cupfunc.R
\name{project_basis}
\alias{project_basis}
\title{This function projects an aligned trait onto the basis
\code{project_basis} projects an external trait into basis space}
\usage{
project_basis(gwas.DT, shrink.DT, pc, traitname = "test_trait")
}
\arguments{
\item{gwas.DT}{data.table object representing summary statistics from external GWAS with columns as defined in \code{\link{get_gwas_data}}}

\item{shrink.DT}{data.table object of matching shrinkage estimates returned by \code{\link{compute_shrinkage_metrics}}}

\item{pc}{prcomp object returned by \code{\link{create_basis}}}

\item{traitname}{character label for this trait defaults to 'test_trait'}
}
\value{
a matrix of PC scores for the projection.
}
\description{
This function projects an aligned trait onto the basis
\code{project_basis} projects an external trait into basis space
}
