% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cupfunc.R
\name{project_sparse}
\alias{project_sparse}
\title{A streamlined function to project a trait onto a sparse basis
\code{project_sparse}}
\usage{
project_sparse(beta, seb, pids)
}
\arguments{
\item{beta}{a vector of beta estimates}

\item{seb}{a vector of standard error of the beta estimates}

\item{pids}{a vector of primary identifiers for SNPs with the same order as beta and seb}
}
\description{
A streamlined function to project a trait onto a sparse basis
\code{project_sparse}
}
\section{Notes}{

This function assumes that the following objects are defined in the current environment
\itemize{
  \item rot.pca - Matrix of rotations, usually obtained from PCA via prcomp.
  \item beta.centers - Vector of basis SNP beta centres, labelled by pid.
  \item shrinkage - Vector of basis SNP shrinkage values, labelled by pid.
  \item LD - Matrix of covariance between basis SNPs
}
This function assumes that the order snps in arguments is the same. Whilst missing SNPs
are allowed this will degrade the projection a warining is issued when more than 5% of SNPs are missing
}

