% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cupfunc.R
\name{se_null}
\alias{se_null}
\title{This function computes standard error under the null
\code{se_null} analytically compute standard error of \eqn{\beta} under \eqn{\mathbb{E}(\beta) = 0}
\eqn{\sqrt{\frac{1}{2fN_0} +  \frac{1}{2N_{0}(1-f)} + \frac{1}{2fN_1} +  \frac{1}{2N_{1}(1-f)} }}}
\usage{
se_null(N, n1, f)
}
\arguments{
\item{N}{a vector or scalar of total number od samples}

\item{n1}{a vector or scalar of number of case samples}

\item{f}{a vector of reference allele frequencies}
}
\value{
a numeric vector
}
\description{
This function computes standard error under the null
\code{se_null} analytically compute standard error of \eqn{\beta} under \eqn{\mathbb{E}(\beta) = 0}
\eqn{\sqrt{\frac{1}{2fN_0} +  \frac{1}{2N_{0}(1-f)} + \frac{1}{2fN_1} +  \frac{1}{2N_{1}(1-f)} }}
}
