% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin2or.R
\name{sigma_e}
\alias{sigma_e}
\title{estimate standard error of epsilon
\code{sigma_e} estimate standard error of epsilon}
\usage{
sigma_e(N, seb, m)
}
\arguments{
\item{N}{sample size}

\item{seb}{- standard error of beta}

\item{m}{- minor allele freq}
}
\value{
a vector of standard error of epsilon
}
\description{
estimate standard error of epsilon
\code{sigma_e} estimate standard error of epsilon
}
