% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_beta}
\alias{simulate_beta}
\title{simulate beta
\code{simulate_beta} use the multivariate normal to simulate realistic beta's}
\usage{
simulate_beta(sm, lor, se_lor, lor_shrink = 1, n_sims = 1)
}
\arguments{
\item{sm}{a snpMatrix object -}

\item{lor}{a vector - observed betas}

\item{se_lor}{a vector - standard error of betas}

\item{lor_shrink}{a vector - shrinkage values to use to adjust betas (default 1 no shrinkage)}

\item{n_sims}{- number of simulations to conduct}
}
\value{
a matrix of simulated betas
}
\description{
simulate beta
\code{simulate_beta} use the multivariate normal to simulate realistic beta's
}
