% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin2or.R
\name{ssx}
\alias{ssx}
\title{estimate allelic sum of squares
\code{ssx} estimate allelic sum of squares}
\usage{
ssx(N, m)
}
\arguments{
\item{N}{sample size}

\item{m}{minor allele freq}
}
\value{
a vector of ssx
}
\description{
estimate allelic sum of squares
\code{ssx} estimate allelic sum of squares
}
