% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/burren_imd_13_traits.R
\docType{data}
\name{stats}
\alias{stats}
\title{Burren sparse basis files}
\format{TODO}
\usage{
data(burren_imd_13_traits)
}
\description{
Burren sparse basis files
}
\examples{
data(burren_imd_13_traits)
}
\references{
TODO
}
\keyword{datasets}
