% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin2or.R
\name{threshBeta}
\alias{threshBeta}
\title{compute OR associated with a given linear coefficient and threshold value
\code{threshBeta}}
\usage{
threshBeta(thresh, N, b, seb, m)
}
\arguments{
\item{thresh}{- threshold value to use.}

\item{N}{sample size}

\item{seb}{- standard error of beta}

\item{m}{- minor allele freq}
}
\value{
a vector of log(OR)
}
\description{
compute OR associated with a given linear coefficient and threshold value
\code{threshBeta}
}
