% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin2or.R
\name{threshOdds}
\alias{threshOdds}
\title{estimate P(Y>thresh|X=x) on odds scale
\code{threshOdds} estimate the probability that Y exceeds a threshold value given an alleleic configuration on the odds scale}
\usage{
threshOdds(thresh, X, N, b, seb, m)
}
\arguments{
\item{thresh}{- threshold value to use.}

\item{X}{(allelic configuration) (0,1,2) assuming additive model}

\item{N}{sample size}

\item{seb}{- standard error of beta}

\item{m}{- minor allele freq}
}
\value{
a vector of odds
}
\description{
estimate P(Y>thresh|X=x) on odds scale
\code{threshOdds} estimate the probability that Y exceeds a threshold value given an alleleic configuration on the odds scale
}
