% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin2or.R
\name{threshSigmaBeta}
\alias{threshSigmaBeta}
\title{compute standard error associated with Beta
\code{threshSigmaBeta}}
\usage{
threshSigmaBeta(thresh, n1, n0, m, tb)
}
\arguments{
\item{thresh}{- threshold value to use.}

\item{n1}{- number of cases}

\item{n0}{- number of controls}

\item{m}{- minor allele freq}

\item{tb}{- number of controls}
}
\value{
a vector standard errors
}
\description{
compute standard error associated with Beta
\code{threshSigmaBeta}
}
