% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{varWeightedEucledian}
\alias{varWeightedEucledian}
\title{helper function to compute variance weighted eucledian distance
\code{varWeightedEucledian} sums logs without loss of precision}
\usage{
varWeightedEucledian(vexp, a, b)
}
\arguments{
\item{vexp}{a numeric vector of variance explained for a set of principle components (e.g summary(pc)$importance[2,])}

\item{a}{a vector of principle component loadings}

\item{b}{a vector of principle component loadings}
}
\value{
a scalar
}
\description{
helper function to compute variance weighted eucledian distance
\code{varWeightedEucledian} sums logs without loss of precision
}
