% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{vcf2snpmatrix}
\alias{vcf2snpmatrix}
\title{convert a vcf file to snpMatrix object
\code{vcf2snpmatrix} convert a vcf file to snpMatrix object}
\usage{
vcf2snpmatrix(vcf, bcf_tools, region_file, quiet = TRUE)
}
\arguments{
\item{vcf}{a scalar - path to vcf file}

\item{bcf_tools}{a scalar - path to bcftools binary}

\item{region_file}{a scalar - path to a file satisfying -R bcftools criteria (optional)}

\item{quiet}{a boolean - if set to false then debug information shown (default = FALSE)}
}
\value{
a list with two slots. sm is the snpMatrix object and info is a data.table describes SNPs
}
\description{
convert a vcf file to snpMatrix object
\code{vcf2snpmatrix} convert a vcf file to snpMatrix object
}
