% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin2or.R
\name{y_bar}
\alias{y_bar}
\title{compute expected value of outcome variable Y
\code{y_bar} compute expected count of ref alleles for a given sample}
\usage{
y_bar(b, N, m)
}
\arguments{
\item{b}{- beta estimate}

\item{N}{sample size}

\item{m}{minor allele freq}
}
\value{
a vector of expected outcome values for Y
}
\description{
compute expected value of outcome variable Y
\code{y_bar} compute expected count of ref alleles for a given sample
}
