% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cupfunc.R
\name{z2p}
\alias{z2p}
\title{convert z to p value
\code{p2z} z to p value}
\usage{
z2p(z)
}
\arguments{
\item{z}{a vector of Z scores}
}
\value{
a vector of p values
}
\description{
convert z to p value
\code{p2z} z to p value
}
