% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{cov_beta}
\alias{cov_beta}
\title{covariance matrix of betas for an ld block
\code{cov_beta}}
\usage{
cov_beta(sm, se_lor)
}
\arguments{
\item{sm}{a snpMatrix object}

\item{se_lor}{a vector - standard error of betas}
}
\value{
a covariance matrix for beta
}
\description{
covariance matrix of betas for an ld block
\code{cov_beta}
}
