% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cupfunc.R
\name{get_gwas_data}
\alias{get_gwas_data}
\title{\code{get_gwas_data} integrate GWAS summary data with support files}
\usage{
get_gwas_data(
  trait_manifest_file,
  snp_manifest_file,
  data_dir,
  filter_snps_by_manifest = FALSE
)
}
\arguments{
\item{trait_manifest_file}{character vector file path to GWAS manifest file}

\item{snp_manifest_file}{character vector file path to snp manifest}

\item{data_dir}{character vector file path to location of GWAS summary stats}

\item{filter_snps_by_manifest}{boolean - whether to prefilter the by snp manifest.
This should be true if you wish to take a subset of SNPs}
}
\value{
data.table object
}
\description{
\code{get_gwas_data} integrate GWAS summary data with support files
}
