% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cupfunc.R
\name{ws_shrinkage}
\alias{ws_shrinkage}
\title{This function computes an alternative to the Bayesian shrinkage method which can be too agressive.
\code{ws_shrinkage} computes a shrinkage based on a weighted sum (ws) of posteriors for each disease
this is then normalised by the total posterior for a given LD block}
\usage{
ws_shrinkage(DT)
}
\arguments{
\item{DT}{basis data.table object}

\item{tquant}{a scalar representing quantile on which to truncate infinite Bayes Factors (DEFAULT 0.9999)}
}
\value{
data.table object
}
\description{
This function computes an alternative to the Bayesian shrinkage method which can be too agressive.
\code{ws_shrinkage} computes a shrinkage based on a weighted sum (ws) of posteriors for each disease
this is then normalised by the total posterior for a given LD block
}
